+++ test/parse.test.js
@@ -20,14 +20,14 @@ describe('connection-parse()', function () {
     expect(res.regular).to.be.a('array');
 
     Object.keys(res.weights).forEach(function (server) {
-      expect(res.weights[server]).to.equal(1);
+      expect(res.weights[server]).to.equal(0);
     });
 
     res.servers.forEach(function (server) {
       expect(server.host).to.be.a('string');
       expect(server.port).to.be.a('number');
       expect(server.string).to.be.a('string');
-      expect(server.weight).to.equal(1);
+      expect(server.weight).to.equal(0);
     });
 
     res.regular.forEach(function (server) {
@@ -43,14 +43,14 @@ describe('connection-parse()', function () {
     expect(res.regular).to.be.a('array');
 
     Object.keys(res.weights).forEach(function (server) {
-      expect(res.weights[server]).to.equal(1);
+      expect(res.weights[server]).to.equal(0);
     });
 
     res.servers.forEach(function (server) {
       expect(server.host).to.be.a('string');
       expect(server.port).to.be.a('number');
       expect(server.string).to.be.a('string');
-      expect(server.weight).to.equal(1);
+      expect(server.weight).to.equal(0);
     });
 
     res.regular.forEach(function (server) {
@@ -109,7 +109,7 @@ describe('connection-parse()', function () {
 
     res = parse({ '1.1.1.1:1111': { foo: 100 }});
     res.servers.forEach(function (server) {
-      expect(server.weight).to.equal(1);
+      expect(server.weight).to.equal(0);
     });
   });
 
@@ -121,14 +121,14 @@ describe('connection-parse()', function () {
     expect(res.regular).to.be.a('array');
 
     Object.keys(res.weights).forEach(function (server) {
-      expect(res.weights[server]).to.equal(1);
+      expect(res.weights[server]).to.equal(0);
     });
 
     res.servers.forEach(function (server) {
       expect(server.port).to.equal(1111);
       expect(server.string).to.equal('1.1.1.1:1111');
       expect(server.host).to.equal('1.1.1.1');
-      expect(server.weight).to.equal(1);
+      expect(server.weight).to.equal(0);
     });
 
     res.regular.forEach(function (server) {
@@ -151,7 +151,7 @@ describe('connection-parse()', function () {
       expect(server.port).to.equal(1111);
       expect(server.string).to.equal('1.1.1.1:1111');
       expect(server.host).to.equal('1.1.1.1');
-      expect(server.weight).to.equal(1);
+      expect(server.weight).to.equal(0);
       expect(server.vnode).to.equal(11);
     });
   });